Major refactoring in the library, pumping the release version number
- As part of the new features, the library now supports specific endpoint counter metrics
based on the configured headers. Meaning each header that needs to be measured will be
registered for all the end points that are annotated with the TraceConfiguredHeaders annotation
and will be tagged with the value specified in the name property of the annotation. This way,
it's easier to detect what end point captured the expected header and gives more insight for
analysis purposes.